Make all numeric types immutable
Small (~2.5%) regression on numeric operations but ~1% improvement on
the OrbitTester benchmark. So it seems like making these mutable did not
have a meaningful impact on performance. We can always do more intense
benchmarking / profiling down the road, but for now this is definitely
the more idiomatic way of working with types that act like numbers.