Adopt Gitflow as the standard workflow for libbgs
This commit is the boundary after which gitflow will be the standard for
adoption of versioning numbers. From this commit forward, any program or
crate using libbgs must use the main branch of libbgs, which represents
a version of libbgs uniquely determined by the version number listed as
a dependency of that program. Semantic versioning has been somewhat
followed since 6cfb974, but the lack of a clear boundary between the dev
and main branches meant that there was ambiguity in the commit being
referenced by a version.

In a development environment, a commit ID should always be used as the
unique identifier of a particular version, but as this research project
begins to transition into data generation, guarantees about the versions
being used must be stronger and easier to follow.

One more historical note: Previously, the main branch was update via the
command `git merge -s ort -Xtheirs dev`, causing commit information to
be lost on the main branch and the main and dev branches to diverge.
This commit was made directly to the dev branch, and the main branch
changed to point to this commit. This commit will be the closest common
parent of all branches moving forward.